+++ packages/core/src/lib/select/index.ts
@@ -1,5 +1,5 @@
 export { SortOptions, type SortOption } from './search';
-export { default as SearchableSelect } from './searchable-select.svelte';
+export { default as TypeaheadInput } from './typeahead-input.svelte';
 export { default as Multiselect } from './multiselect.svelte';
 export { default as Select } from './select.svelte';
 export { default as SelectInput } from './select-input.svelte';
